Build response for Terraform Workflow (#607)
Now that the conftest activity returns the set of failed policies, we
can feed it through as a response of the Terraform Worfklow. Note that
the parent Deploy workflow does not generate/attempt to process these
failed policies since platform mode doesn't run the validate step. We
will only process it when the parent workflow is the PR, which will be
the next major change.