Support atlantis apply on neptune for teams. (#617)
## Summary

Introduces the concept of deploy requirements and adds two: 
Team - user must be in the appropriate gh team, we skip if there is no
team configured in the apply settings
Branch Restriction - branch restriction setting must be `none` in order
to proceed. This ensures that we don't just open up `atlantis apply` to
the world

There's some additional changes here such as a new PR template which
will be turned into a comment if there's any error whatsover in the
async portion of the gateway request. I think it's reasonable to have
this seeing as the current functionality is to provide no feedback and
just drop a log. This makes our system seem unreliable as the customer
thinks atlantis isn't doing anything.

We can evaluate how this works and if we need to add a status to block
the PR for certain cases we can. Right now, since `atlantis apply`
comments are necessary anymore anyways, a comment should suffice.


**Note docker desktop is now blocked on my computer it seems so I can't
test this locally, we'll need to figure out a new way to test things
locally.**

## TODO 

- [x] Requirement unit testing
- [x] Staging Testing