Modify where project workflow mode type is overriden. (#621)
This was initially to ensure that passing "pr" wouldn't fail since
that's what we expect for the global key.

But i don't think we should be overriding project config when we are
merging the repo config with the global config. Repo config should be
created with the defaults/values at that point already.