Signal PR workflow on both approval and changes_requested pr review events (#689)
Gateway changes: We are going to want to update our check run status
when a `changes_requested` event occurs to be able to support failing
check runs should a reviewer change their mind and want to request
changes subsequently after an initial approval.

PR workflow changes: Renaming of approval objects to use "review"
instead. We will be restructuring the code to support this edge case in
subsequent PRs.