Avoid failing child TF workflow on validation failure (#692)
Initially we failed the TF workflow if the validation step returned
failed Conftest policies for the corresponding root (this allowed us the
easily parse out that error to perform the correct update logic for our
GH check run). Bug testing seems to hint that if we do fail the workflow
(i.e. return an error), we don't return any workflow result either.
Solution here is to simply mark the workflow as successful even if the
Terraform plan doesn't pass all validation requirements. We can process
the results itself to set the failure reason to
`ValidationFailedReason`.

This better lines up with my interpretation of a successful TF workflow
since failing validation checks shouldn't mean the TF workflow didn't do
its job correctly (we do the same behavior on the Conftest activity
itself). Plus, it avoids having to create a more complex error type to
contain all of the metadata we want returned to the parent PR workflow.