Update check run after all failing policies are approved (#693)
We support using GH approvals as a way to bypass failing policies. If
all failing policies on a root are approved, we need to mark the
checkrun status as successful to allow for merges. Note this is
triggered by approval events (we will rerun full plan and policy check
steps if a review = requested changes to revert back to no approvals).