Fix team list fetch to use Name identifier (#696)
Bug fix in approval logic: Make sure we fetch the set of owners on a
team using the team name (i.e. `teams[failedPolicy.Name]`) instead of
owners because that is the key we build the list off of. PR also goes
through and consistently names the remaining failed policies to the
variable: `remainingFailedPolicies`.