Always update policy check runs in legacy regardless of feature flag (#706)
Originally, I gated calls to create/update check runs in legacy with the
runtime feature flag Legacy-Deprecation, to ensure we only mutate check
runs when the PR workflow isn't tasked to do so (and vice versa).
However, the PR workflow does not ever handle the policy check run
(which we will eventually delete when rollout is complete) since we've
merged them together with the plan check run.

Unfortunately, to maintain this state and ability to rollout and
rollback easily, we need to always create and update the policy check
run, regardless of whether or not the new workflow is managing the PR
lifecycle. That way, existing PRs using the legacy system won't get
stuck on a failing policy check run when we migrate over to the PR
workflow.