Avoid adding invalid check runs to the check run cache (#708)
The problem with enabling a feature allocator at the check run mutation
level is the temporal worker relies on a built in cache to store in
progress check run IDs. This can cause issues down the line because if
we are now tasked with updating this check run after enabling the
feature flag, the original logic will store an empty check run id in the
cache, causing us to attempt to Update the check run with a "0" id.

With this change, I want test out skipping storing those 0 id check runs
and see what the UX behavior is if we just create a new check run to
replace any existing ones.