Support retries on first commit (#713)
We originally required retries to only be allowed on the same revision
as the last attempted deploy to prevent accidental reverts in history.
However, there is an edge case where the first deployment on a root
needs a retry. When it's the first deployment, the commit direction will
always default to `DirectionAhead` instead of `DirectionIdentical` so we
need to handle that scenario separately.

This change is technically a non-deterministic change (very unlikely to
occur, as it's a rare scenario) so I've protected the rollout with
versioning. While doing this, I also went ahead and cleaned up old
versions in the Deploy workflow that are no longer relevant.