Check if key is in roots map of Processor (#718)
Before notifying, we should ensure the key is in the roots map to
prevent attempts to notify check run on invalid, empty repo metadata.
There is a chance the root map can miss a key if multiple revisions come
in quickly and a new Processor for the latest revision is receiving a
signal on an old revision. We shouldn't care about this signal since we
already perform a clean defer to cancel those check runs on shutdown of
the old revision Processor.